Add GameRunner and MarioGame implementations
A new class, GameRunner, was added along with the MarioGame class. The GameRunner accepts a MarioGame object through its constructor and runs the introduced game. This change was made for setting up a basic framework for the gaming application that enables run any game, starting with Mario Game. The main AppGamingBasic class was also modified to create a MarioGame instance and run it